+++ dom/media/mediasource/SourceBufferResource.h
@@ -113,7 +113,7 @@ public:
   }
   
   
-  uint32_t EvictData(uint64_t aPlaybackOffset, uint32_t aThreshold,
+  uint32_t EvictData(uint64_t aPlaybackOffset, int64_t aThreshold,
                      ErrorResult& aRv);
